fix(params): rename CALIBRE_AUTH to CALIBRE_API_TOKEN
