+++ .circleci/config.yml
@@ -55,45 +55,23 @@ jobs:
     - run:
         name: revert changes to package-lock.json
         command: git checkout -- package-lock.json
-
     - run:
         name: configure wsk
         command: echo -e "APIHOST=https://adobeioruntime.net\nNAMESPACE=helix\nAUTH=$WSK_AUTH\n" > ~/.wskprops
-
     - run:
         name: configure env
         command: echo -e "LOG_LEVEL=info\n" > .env
-
     - run:
         name: Semantic Release
         command: npm run semantic-release
-
-    - run:
-        name: Set Google Token
-        command: wsk action update helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@`node -e "console.log(require('./package.json').version)"` --timeout 300000 -p CALIBRE_AUTH $CALIBRE_AUTH
-
     - run:
         name: Post-Deployment Integration Test
         command: npm run test-postdeploy
-
     - store_test_results:
         path: junit
-
     - run:
-        name: Set Epsagon Token and Google Token
-        command: wsk action update helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@`node -e "console.log(require('./package.json').version)"` -p EPSAGON_TOKEN $EPSAGON_TOKEN  --timeout 300000 -p CALIBRE_AUTH $CALIBRE_AUTH
-
-    - run:
-        name: Link Sequence (Latest)
-        command: wsk action update --sequence /helix/helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@latest helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@`node -e "console.log(require('./package.json').version)"` --web=true --web-secure=false
-
-    - run:
-        name: Link Sequence (Major)
-        command: wsk action update --sequence /helix/helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@v`node -e "console.log(require('./package.json').version.match(/^[0-9]+/)[0])"` helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@`node -e "console.log(require('./package.json').version)"` --web=true --web-secure=false
-
-    - run:
-        name: Link Sequence (Minor)
-        command: wsk action update --sequence /helix/helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@v`node -e "console.log(require('./package.json').version.match(/^[0-9]+.[0-9]+/)[0])"` helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@`node -e "console.log(require('./package.json').version)"` --web=true --web-secure=false
+        name: Link Sequences
+        command: npm run deploy-sequences
 
   branch-deploy:
     executor: node10
@@ -103,33 +81,15 @@ jobs:
     - run:
         name: revert changes to package-lock.json
         command: git checkout -- package-lock.json
-
     - run:
         name: configure wsk
         command: echo -e "APIHOST=https://adobeioruntime.net\nNAMESPACE=helix\nAUTH=$WSK_AUTH\n" > ~/.wskprops
-
     - run:
         name: Branch Deployment
-        command: npx wsk-builder -v --entryFile index.js --deploy --test --name helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@ci$CIRCLE_BUILD_NUM
-
-    - run:
-        name: Set Google Token
-        command: wsk action update helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@ci$CIRCLE_BUILD_NUM -p  --timeout 300000 -p CALIBRE_AUTH $CALIBRE_AUTH
-
+        command: npm run deploy-ci
     - run:
         name: Post-Deployment Integration Test
         command: npm run test-postdeploy
-        
-    - run:
-        name: Set Epsagon Token and Google Token
-        command: wsk action update helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@ci$CIRCLE_BUILD_NUM -p EPSAGON_TOKEN $EPSAGON_TOKEN   --timeout 300000 -p CALIBRE_AUTH $CALIBRE_AUTH
-
-    - store_test_results:
-        path: junit
-
-    - run:
-        name: Link Sequence (CI)
-        command: wsk action update --sequence /helix/helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@ci helix-services/`node -e "console.log(require('./package.json').name.replace('@adobe/helix-', ''))"`@ci$CIRCLE_BUILD_NUM --web=true --web-secure=false
 
 workflows:
   version: 2
