+++ .eslintignore
@@ -1,2 +1,4 @@
 .vscode/*
 coverage/*
+dist
+node_modules
