+++ .gitignore
@@ -6,3 +6,4 @@ tmp
 logs
 .DS_Store
 test-results.xml
+dist
