+++ package.json
@@ -3,15 +3,17 @@
   "version": "1.0.3",
   "description": "Helix Performance",
   "main": "index.js",
+  "private": true,
   "scripts": {
     "test": "nyc mocha -i -g Post-Deployment --exit",
     "test-ci": "npm run lint && nyc --reporter=text --reporter=lcov --check-coverage --branches 90 --statements 90 --lines 90 mocha --reporter xunit --reporter-options output=./junit/test-results.xml -i -g Post-Deployment  --exit && codecov",
     "test-postdeploy": "mocha --reporter xunit --reporter-options output=./junit/test-results.xml -g Post-Deployment  --exit",
     "semantic-release": "semantic-release",
-    "semantic-prepare": "node ./build/prepare.js",
+    "deploy-sequences": "wsk-builder --no-build -no-hints -l latest -l major -l minor",
+    "deploy-ci": "wsk-builder -v --deploy --test=/_status_check/pingdom.xml --pkgVersion=ci$CIRCLE_BUILD_NUM -l ci",
     "lint": "./node_modules/.bin/eslint .",
     "build": "wsk-builder -v --entryFile index.js",
-    "deploy": "wsk-builder -v --deploy --test --entryFile index.js -p EPSAGON_TOKEN=$EPSAGON_TOKEN",
+    "deploy": "wsk-builder -v --deploy --test=/_status_check/pingdom.xml",
     "snyk-protect": "snyk protect",
     "prepare": "npm run snyk-protect"
   },
@@ -26,42 +28,42 @@
   },
   "homepage": "https://github.com/adobe/helix-perf#readme",
   "dependencies": {
-    "@adobe/fastly-native-promises": "^1.10.0",
-    "@adobe/helix-pingdom-status": "^3.0.0",
-    "@adobe/openwhisk-action-builder": "^1.1.1",
-    "calibre": "^2.2.1",
-    "epsagon": "^1.29.0",
-    "snyk": "^1.192.6",
+    "@adobe/fastly-native-promises": "^1.11.0",
+    "@adobe/helix-pingdom-status": "^3.1.0",
+    "@adobe/openwhisk-action-builder": "^1.3.1",
+    "calibre": "^2.3.2",
+    "epsagon": "^1.31.7",
+    "snyk": "^1.216.0",
     "uri-js": "^4.2.2"
   },
   "devDependencies": {
-    "@adobe/helix-shared": "^2.0.0",
+    "@adobe/helix-shared": "^2.1.1",
     "@semantic-release/changelog": "^3.0.4",
     "@semantic-release/commit-analyzer": "^6.2.0",
     "@semantic-release/exec": "^3.3.5",
     "@semantic-release/git": "^7.0.16",
     "@semantic-release/github": "^5.4.2",
     "@semantic-release/release-notes-generator": "^7.2.1",
-    "ajv": "^6.10.1",
+    "ajv": "^6.10.2",
     "codecov": "^3.5.0",
-    "eslint": "^6.0.1",
+    "eslint": "^6.1.0",
     "eslint-config-airbnb": "^17.1.1",
-    "eslint-plugin-header": "^2.0.0",
-    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-header": "^3.0.0",
+    "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.14.2",
+    "eslint-plugin-react": "^7.14.3",
     "istanbul": "^1.1.0-alpha.1",
-    "junit-report-builder": "^1.3.2",
-    "mocha": "^6.1.4",
-    "mocha-junit-reporter": "^1.23.0",
+    "junit-report-builder": "^1.3.3",
+    "mocha": "^6.2.0",
+    "mocha-junit-reporter": "^1.23.1",
     "mocha-parallel-tests": "^2.2.1",
     "nyc": "^14.1.1",
-    "proxyquire": "^2.1.0",
-    "semantic-release": "^15.13.18",
-    "sinon": "^7.3.2"
+    "proxyquire": "^2.1.2",
+    "semantic-release": "^15.13.19",
+    "sinon": "^7.4.1"
   },
   "snyk": true,
   "wsk": {
-    "name": "helix-services/perf@1.0.3"
+    "name": "helix-services/perf@${version}"
   }
 }
